requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1

# GitHub API
PyGithub==2.1.1
requests==2.31.0

# AI Models
google-generativeai==0.3.2
openai==1.3.7

# Additional utilities
python-dotenv==1.0.0
schedule==1.2.0
cryptography==41.0.7
pydantic==2.5.2